feat: skip if no settings

+++ src/migrations/20250613171510-sync-locks.ts
@@ -37,6 +37,8 @@ export const lockMigration: IMigration = {
         }),
       )
 
+      if (settings.length === 0) return
+
       // remove all old locks
       await Models.Lock.deleteMany({})
